Refactor submission form button
An issue was noticed when submitting. When pressing enter the form
submitted instantly and by-passed the required fields, and also,
whether event.preventDefault() was called on the submit button
seemed to have no difference whatsoever.

It was then learnt that submit event is fired on the form itself,
so the event listener was changed to match that. The order of the
submit buttons was also changed, so now when pressing Enter it will
use add Book button, which causes the browser's validation to corr-
ectly work and as such asks you to fill the remaining fields first.