Fix cancel button bug
Present was a bug with the cancel button. When clicked it would
add a new empty book to the library. The reason as to why, is
because the form submit event gets fired on any submit button
so we had to change the type of the button to button and then
add a listener so that if clicked the modal dialog closes.