Remove read form input
If adding a book to a library, one would assume that the person has
not read the book, as the default. Then the person can change the
status of the book once they have finished reading it. Therefore
it seemed a bit unneeded to have a form input that asks the user
if they had read the book or not.